src/tests/test_trader.py
import pytest

def test_trader_initialization(test_trader):
    """Test the initialization of the Trader object."""
    assert test_trader is not None, "Trader should be initialized."

def test_buy_signal(test_trader, test_strategy, mocker):
    """Test processing of buy signal."""
    mocker.patch.object(test_strategy, 'should_buy', return_value=True)
    mocker.patch.object(test_trader, 'execute_trade', return_value=None)
    test_trader.run()  # Implement run to handle one iteration for testability
    test_strategy.should_buy.assert_called_once()
    test_trader.execute_trade.assert_called_with('buy')

def test_sell_signal(test_trader, test_strategy, mocker):
    """Test processing of sell signal."""
    mocker.patch.object(test_strategy, 'should_sell', return_value=True)
    mocker.patch.object(test_trader, 'execute_trade', return_value=None)
    test_trader.run()  # Similar modification for testability
    test_strategy.should_sell.assert_called_once()
    test_trader.execute_trade.assert_called_with('sell')
